requirements.txt
opencv-contrib-python>=4.5.0
customtkinter>=5.0.0
Pillow>=9.0.0
numpy>=1.21.0
pymongo>=4.0.0
dnspython>=2.0.0
requests>=2.25.0
deepface>=0.0.79
tensorflow>=2.10.0
tkcalendar>=1.6.0
openpyxl>=3.0.0
